Fix cyclic dependency between MC and DebugInfoDWARF when building clang with modules enabled
