+++ llvm/include/llvm/MC/CAS/MCCASObjectV1.h
@@ -12,11 +12,6 @@
 #include "llvm/BinaryFormat/MachO.h"
 #include "llvm/CAS/CASID.h"
 #include "llvm/CAS/ObjectStore.h"
-#include "llvm/DebugInfo/DWARF/DWARFCompileUnit.h"
-#include "llvm/DebugInfo/DWARF/DWARFContext.h"
-#include "llvm/DebugInfo/DWARF/DWARFDataExtractor.h"
-#include "llvm/DebugInfo/DWARF/DWARFDebugAbbrev.h"
-#include "llvm/DebugInfo/DWARF/DWARFDebugLine.h"
 #include "llvm/MC/CAS/MCCASFormatSchemaBase.h"
 #include "llvm/MC/CAS/MCCASReader.h"
 #include "llvm/MC/MCAsmLayout.h"
@@ -26,26 +21,6 @@
 
 namespace llvm {
 
-template <> struct DenseMapInfo<llvm::dwarf::Form> {
-  static llvm::dwarf::Form getEmptyKey() {
-    return static_cast<llvm::dwarf::Form>(
-        DenseMapInfo<uint16_t>::getEmptyKey());
-  }
-
-  static llvm::dwarf::Form getTombstoneKey() {
-    return static_cast<llvm::dwarf::Form>(
-        DenseMapInfo<uint16_t>::getTombstoneKey());
-  }
-
-  static unsigned getHashValue(const llvm::dwarf::Form &OVal) {
-    return DenseMapInfo<uint16_t>::getHashValue(OVal);
-  }
-
-  static bool isEqual(const llvm::dwarf::Form &LHS,
-                      const llvm::dwarf::Form &RHS) {
-    return LHS == RHS;
-  }
-};
 namespace mccasformats {
 namespace v1 {
 
@@ -532,48 +507,6 @@ public:
 private:
   const Triple &Target;
   const MCSchema &Schema;
-};
-
-/// A DWARFObject implementation that can be used to dwarfdump CAS-formatted
-/// debug info.
-class InMemoryCASDWARFObject : public DWARFObject {
-  ArrayRef<char> DebugAbbrevSection;
-  bool IsLittleEndian;
-
-public:
-  InMemoryCASDWARFObject(ArrayRef<char> AbbrevContents, bool IsLittleEndian)
-      : DebugAbbrevSection(AbbrevContents), IsLittleEndian(IsLittleEndian) {}
-  bool isLittleEndian() const override { return IsLittleEndian; }
-
-  StringRef getAbbrevSection() const override {
-    return toStringRef(DebugAbbrevSection);
-  }
-
-  Optional<RelocAddrEntry> find(const DWARFSection &Sec,
-                                uint64_t Pos) const override {
-    return {};
-  }
-
-  /// This struct represents the Data in one Compile Unit. The DistinctData is
-  /// the data that doesn't deduplicate and must be stored separately, the
-  /// DebugInfoRefData is the data that is stored in one DebugInfoCURef cas
-  /// object and will deduplicate for a link ODR function.
-  struct PartitionedDebugInfoSection {
-    SmallVector<char, 0> DebugInfoCURefData;
-    SmallVector<char, 0> DistinctData;
-    constexpr static std::array FormsToPartition{
-        llvm::dwarf::Form::DW_FORM_strp, llvm::dwarf::Form::DW_FORM_sec_offset};
-  };
-
-  /// Create a DwarfCompileUnit that represents the compile unit at \p CUOffset
-  /// in the debug info section, and iterate over the individual DIEs to
-  /// identify and separate the Forms that do not deduplicate in
-  /// PartitionedDebugInfoSection::FormsToPartition and those that do
-  /// deduplicate. Store both kinds of Forms in their own buffers per compile
-  /// unit.
-  Expected<PartitionedDebugInfoSection>
-  splitUpCUData(ArrayRef<char> DebugInfoData, uint64_t AbbrevOffset,
-                uint64_t CUOffset, DWARFContext *Ctx);
 };
 
 class DebugInfoSectionRef : public SpecificRef<DebugInfoSectionRef> {
